astro: share content config
